fix: retain feed page size
Currently when getting the next page of a feed, we clear the page size
that it was created with. This means that if I start a feed with page
size 1000, after that my subsequent pages will drop to the default page
size of 16. This change makes it such that the page size will stay
consistent.